Adjust Swift tools version to 5.5 to work with Xcode on Big Sur.
Unfortunately, Xcode versions 13.1, 13.2, and 13.2.1 didn't increment the version of SwiftPM,
despite incrementing the Swift version to 5.5.1, and 5.5.2, respectively.

Since the next release of Xcode, 13.3, requires macOS 12 (Monterey), there's no released version
of Xcode for macOS 11 (Big Sur) that can parse a Package.swift with swift-tools-version:5.5.2.

SwiftPM 5.5.0 also refuses to process manifests with SwiftVersion.version("5.5.2"), so we have
to rely on the compiler itself to require Swift 5.5.2.